requirements.txt
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
modeltranslation==0.25
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
requests==2.32.2
rfc3986==1.5.0
sniffio==1.3.1
urllib3==2.2.1
